[refactor] carry Batch.data_timestamp in seconds (float), not ms
Make the event-time feature single-unit. Batch.data_timestamp is now
Unix-epoch seconds (float), matching the config (interval/targets in seconds),
the persisted watermark, the trigger math, and Python time.time(). Normalization
from the raw kafka ms happens once in _build_batch (max_ts / 1000.0); the
internal __data_timestamp__ arrow column keeps the raw ms. The all_gather no
longer needs a /1000 conversion. float64 preserves full ms precision at
epoch-seconds magnitude.

+++ tzrec/datasets/kafka_dataset_test.py
@@ -308,9 +308,9 @@ class KafkaDatasetTest(unittest.TestCase):
             self.assertGreaterEqual(value, 0)
 
         # Event-time should be surfaced from the kafka message timestamps (the
-        # producer stamps create-time), as a positive epoch-ms integer.
+        # producer stamps create-time), as positive Unix-epoch seconds (float).
         self.assertIsNotNone(batch.data_timestamp)
-        self.assertIsInstance(batch.data_timestamp, int)
+        self.assertIsInstance(batch.data_timestamp, float)
         self.assertGreater(batch.data_timestamp, 0)
 
     @unittest.skipIf(
